Separate libraries for parsers (tests does not work)
