+++ client/lib/CMakeLists.txt
@@ -1,0 +1,2 @@
+add_subdirectory(connection)
+add_subdirectory(plugin_wrappers)
