+++ client/lib/plugin_wrappers/audio/AudioPluginWrapper.cpp
@@ -1,0 +1,1 @@
+#include "AudioPluginWrapper.h"
