+++ client/lib/plugin_wrappers/basic/CMakeLists.txt
@@ -1,0 +1,11 @@
+add_library(basic_plugin_wrapper BasicPluginWrapper.cpp)
+
+add_executable(basic_plugin_wrapper_test BasicPluginWrapper_tests.cpp)
+
+target_link_libraries(basic_plugin_wrapper PUBLIC connection)
+target_link_libraries(basic_plugin_wrapper_test Private GTest::gtest_main)
+
+target_include_directories(basic_plugin_wrapper PUBLIC ./)
+target_include_directories(basic_plugin_wrapper PUBLIC ../include)
+
+add_test(basic_plugin_wrapper_test basic_plugin_wrapper_test)
