+++ client/lib/plugin_wrappers/basic/IBasicPluginWrapper.h
