+++ client/lib/plugin_wrappers/format_processor/FormatProcessorPluginWrapper.cpp
