+++ client/lib/plugin_wrappers/image/IImagePluginWrapper.h
@@ -4,7 +4,7 @@
 #include <string>
 #include <vector>
 
-#include "interfaces/IBasicPluginWrapper.h"
+#include "../basic/IBasicPluginWrapper.h"
 
 struct IImagePluginWrapper : virtual public IBasicPluginWrapper {
     virtual std::pair<std::vector<std::string>, std::string>
