+++ client/lib/plugin_wrappers/image/ImagePluginWrapper.cpp
