+++ client/lib/plugin_wrappers/image/ImagePluginWrapper_tests.cpp
