+++ client/lib/plugin_wrappers/sentence/SentencePluginWrapper.cpp
